Separated the commands into subdirectories
